pyproject.toml
[tool.poetry]
name = "fitbit-obsidian-sync"
version = "0.1.0"
description = ""
authors = ["Matt <mwalczyk81@gmail.com>"]
readme = "README.md"
packages = [{ include = "fitbit_sync", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fitbit = "^0.3.1"
requests = "^2.32.3"
oauthlib = "^3.2.2"
markdownify = "^1.1.0"
python-dotenv = "^1.1.0"
cherrypy = "^18.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
